# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puck Page Builder API",
        "description": "API endpoints for the Puck-based page builder integration",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
        "/api/puck/page": {
            "get": {
                "tags": [
                    "Puck Pages"
                ],
                "summary": "Get page by path",
                "description": "Fetches a page by its URL path for Puck frontend rendering",
                "operationId": "getPageByPath",
                "parameters": [
                    {
                        "name": "path",
                        "in": "query",
                        "description": "URL path of the page (e.g., '/', '/about')",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "/",
                            "example": "/"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "path": {
                                            "type": "string",
                                            "example": "/"
                                        },
                                        "data": {
                                            "properties": {
                                                "root": {
                                                    "type": "object"
                                                },
                                                "content": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "zones": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Puck Pages"
                ],
                "summary": "Create or update page (upsert)",
                "description": "Creates a new page or updates existing one by path. Requires EDITOR or SUPER_ADMIN permissions.",
                "operationId": "savePuckPage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "path",
                                    "title"
                                ],
                                "properties": {
                                    "path": {
                                        "description": "URL path for the page",
                                        "type": "string",
                                        "example": "/"
                                    },
                                    "title": {
                                        "description": "Page title",
                                        "type": "string",
                                        "example": "Homepage"
                                    },
                                    "slug": {
                                        "description": "URL slug (auto-generated if not provided)",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "data": {
                                        "description": "Puck editor data structure",
                                        "properties": {
                                            "root": {
                                                "properties": {
                                                    "props": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "content": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "type": {
                                                            "type": "string",
                                                            "example": "HeroSlider"
                                                        },
                                                        "props": {
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "zones": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "meta": {
                                        "description": "SEO and other metadata",
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "/"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}